+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 ### Changed
 - Integrations - Change the Layer integration to be based on the Server API.
 
-### Fixed
+### Fixed
 - Code Inspection - Fix Forest customization code inspection to be recursive through directories.
 
 ## RELEASE 1.2.1 - 2017-08-23