+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased]
 ### Fixed
-- Exports - Espace special characters for the string fields. 
+- Exports - Escape special characters for the string fields. 
 
 ## RELEASE 1.2.2 - 2017-08-24
 ### Changed